// front/src/api/generated/schemas/region.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Engineer Route Plan API
 * OpenAPI spec version: 0.1.0
 */
import type { RegionCode } from './regionCode';

export interface Region {
  code: RegionCode;
  /** Название региона для показа пользователю */
  name: string;
}